+++ src/collective/iconifiedcategory/browser/static/iconifiedcategory.js
@@ -90,8 +90,9 @@ jQuery(function($) {
 
 };
 
-function categorizedChildsInfos() {
-    tooltipster_helper(selector='.tooltipster-childs-infos',
+function categorizedChildsInfos(options={}) {
+    selector = options.selector || '.tooltipster-childs-infos';
+    tooltipster_helper(selector=selector,
                        view_name='@@categorized-childs-infos',
                        data_parameters=['category_uid']);